+++ code_pydocstyle.py
@@ -24,14 +24,13 @@ MAX_ERROR_BY_TYPE = {
     # If the error code is not in this dict, then there is no tolerance on the error.
     # http://www.pydocstyle.org/en/stable/error_codes.html
     'D101': 50,
-    'D102': 656,
+    'D102': 641,
     'D103': 30,
-
-    'D205': 214,
+    'D205': 200,
 
     'D300': 6,
 
-    'D400': 228,
+    'D400': 245,
     'D403': 44,
 }