+++ tests/faces/test_bsplinesurface3d.py
@@ -16,7 +16,7 @@ class TestBSplineSurface3D(unittest.TestCase):
         grid2d = volmdlr.grid.Grid2D.from_properties((0, 1), (0, 1), (10, 10))
         contour2d_dim = bspline_surfaces.bspline_surface_2.contour2d_parametric_to_dimension(contour2d, grid2d)
         self.assertEqual(len(contour2d_dim.primitives), 4)
-        self.assertAlmostEqual(contour2d_dim.area(), 18.112581823847673, places=2)
+        self.assertAlmostEqual(contour2d_dim.area(), 18.12437118716217, places=2)
         self.assertAlmostEqual(contour2d_dim.length(), 16.81606170335965, places=2)
 
     def test_periodicity(self):
@@ -46,6 +46,23 @@ class TestBSplineSurface3D(unittest.TestCase):
         self.assertIsInstance(inv_prof, vme.Arc3D)
         self.assertEqual(inv_prof.start, arc.start)
         self.assertEqual(inv_prof.end, arc.end)
+
+    def test_bsplinecurve3d_to_2d(self):
+        bspline_surface = bspline_surfaces.bspline_surface_4
+        control_points = [volmdlr.Point3D(-0.012138106431296442, 0.11769707710908962, -0.10360094389690414),
+         volmdlr.Point3D(-0.012153195391844274, 0.1177764571887428, -0.10360691055433219),
+         volmdlr.Point3D(-0.01216612946601426, 0.11785649353385147, -0.10361063821784446),
+         volmdlr.Point3D(-0.012176888504086755, 0.11793706145749239, -0.10361212108019317)]
+        weights = [1.0, 0.9994807070752826, 0.9994807070752826, 1.0]
+        original_bspline = vme.BSplineCurve3D(3, control_points, [4, 4], [0, 1], weights, False)
+        bspline_on_parametric_domain = bspline_surface.bsplinecurve3d_to_2d(original_bspline)[0]
+        bspline_after_transfomation = bspline_surface.bsplinecurve2d_to_3d(bspline_on_parametric_domain)[0]
+        original_length = original_bspline.length()
+        length_after_transformation = bspline_after_transfomation.length()
+        point = original_bspline.point_at_abscissa(0.5*original_length)
+        point_test = bspline_after_transfomation.point_at_abscissa(0.5 * length_after_transformation)
+        self.assertAlmostEqual(original_length, length_after_transformation, places=6)
+        self.assertTrue(point.is_close(point_test, 1e-6))
 
 
 if __name__ == '__main__':