+++ volmdlr/core.py
@@ -78,8 +78,7 @@ def delete_double_point(list_point):
 
 def step_ids_to_str(ids):
     """
-    Returns a string with a '#' in front of each ID and a comma separating
-    eachone.
+    Returns a string with a '#' in front of each ID and a comma separating eachone.
 
     :param ids: A list of step primitives IDs
     :type ids: List[int]
@@ -314,7 +313,7 @@ class Primitive3D(dc.PhysicalObject):
 
 class CompositePrimitive3D(CompositePrimitive, Primitive3D):
     """
-    A collection of simple primitives3D
+    A collection of simple primitives3D.
     """
     _standalone_in_db = True
     _eq_is_data_eq = True
@@ -900,7 +899,8 @@ class VolumeModel(dc.PhysicalObject):
     def frame_mapping_inplace(self, frame: volmdlr.Frame3D, side: str):
         """
         Changes frame_mapping and the object is updated inplace.
-        side = 'old' or 'new'
+
+        side = 'old' or 'new'.
         """
         for primitives in self.primitives:
             primitives.frame_mapping_inplace(frame, side)
@@ -927,6 +927,7 @@ class VolumeModel(dc.PhysicalObject):
     def plot(self, equal_aspect=True):
         """
         Matplotlib plot of model.
+
         To use for debug.
         """
         fig = plt.figure()
@@ -1035,6 +1036,11 @@ class VolumeModel(dc.PhysicalObject):
         return output
 
     def babylon_data(self):
+        """
+        Get babylonjs data.
+
+        :return: Dictionary with babylon data.
+        """
         meshes = []
         lines = []
         for primitive in self.primitives:
@@ -1683,6 +1689,11 @@ class MovingVolumeModel(VolumeModel):
         return VolumeModel(primitives)
 
     def babylon_data(self):
+        """
+        Get babylonjs data.
+
+        :return: Dictionary with babylon data.
+        """
         meshes = []
         primitives_to_meshes = []
         for i_prim, primitive in enumerate(self.primitives):